int-tests: Make it possible to start apim only once while writing tests. Add readme.
int-tests: Remove additional todo. Update readme.

int-tests: suggestions: rename the withapim setup classes

int-tests: Add the missing license

int-tests: Fix failing tests
